Combat Update 3/18/2018
Implimented [locate()] changes with seeming success. I also did a bug fix for flurry to keep it from rollnig 500's, made it so that Adaptability makes a bigger difference to AP but does not go to less than 1. And i think another change or two that I can't quite recall atm.