Combat Command Updates
Not sure what I changed for Charge, i think i was tweaking the balance of the extra damage, but ended up leaving it largely alone.

Drain and Absorb had their damage adjusted up from about 20% to 25%

Feint has been changed to not only have a maximum amount of 1000 for single jutsu being enhanced by it, but I also made it actually work the way it was suppossed to where it can be used to assist other's jutsu and not just your own.

I changed Counter so that it has a lower chance of success without Effectiveness so it isn't all about the Precision.