Combat Command Updates
Adjusted Link to only cost 3 AP instead of 4 for those without Yin affinity.

Made it so that Yank can be used outside of the players turn to better assist in situations of team work against Genjtusu.

Made sure Kai costs 1 AP to use.

Updated Counter Link

Adjusted Feint from a max of 1000 to 750 and made it so that its calculation is based off of the contributing jutsu instead of the current roll of the assisted jutsu.

Adjusted Block and Turtle in resource costs slightly while alleviating their roll penalties to make them competative options versus Evade and Avoid once more.

Adjusted Stamina and Chakra costs of Tags and Scrolls

Fixed Control and Resist to pull the proper data for checks.