Combat Function Update
Sacrifice Expertise adjusted in TenseRoll

Perception Checks now have Extra Sensory integration

Evade and Avoid updated to utilize Evasive Maneuvers Expertise

Block ,Turtle, and Barrier given the ability to absorb the chakra of a jutsu that they successfully fully defend against.

BarrierDur has had Barrier expertise added to it.

All defenses have had an adjustment to the FormDefMod to fix what was going in to it.