Combat Function Update
Form's influence on crit has been adjusted so that Ninjutsu gets less benefit from it than Taijutsu. This is to reflect the fact that Form is far more necessary for Ninjutsu than Taijutsu, and that Ninjutsu hits harder in general. This should help to make Tai more competative with Ninjutsu.

Adjusted some name() to name(locate()).

Fixed Regen's emit to match Regen's effect.