Combat Function Update
Combat Ap made to calculate correctly (generally a decrease i nthe amount of Form needed to min AP, but an increase to the amount of Form and Adaptability to lower Spec Actions)

Adjusted FormDef to be more potent. Raised from -12/-24 to Crit roll to -40 max to Crit roll. This is to make it so that investing in form for Defenses has more value for protecting against crits.

Tense and Sacrifice were adjusted to actually take more damage from Powerful spec actions.

Barrier was changed to make it so that you cannot use it on people who are not actually being attacked (would cause a bug if attempted prior) and it will now properly grow in strength again (hopefully) with more people covered by the barrier.

EAssault damage for General has been raised slightly. Chakra drains for Hyuuga and Aburame have been lowered.