Combat Command Update
Made it so that attacks for the most part record their chakra so that those with Sensitive can see it when they are attacked by jutsu as a tactical advantage.

Made it so that damage is displayed in defendprompt so that players have a better idea what sort of jtusu they are facing and don't die by accident.

Made is so that scroll and tag attacks can be used by staff bits against themselves.

Adjusted some costs for EAssault